app/stream/requirements.txt
schedule
kafka-python
pyspark
requests
feedparser
pyspark==3.5.1
great_expectations
pandas
psycopg2-binary